Lecture 8. watched. and updated android studio

+++ app/src/main/java/com/jonitiainen/edistynytmobiiliohjelmointi/datatypes/weatherstation/_8.java
@@ -10,13 +10,13 @@ public class _8 {
 
     @SerializedName("v")
     @Expose
-    private int v;
+    private double v;
 
-    public int getV() {
+    public double getV() {
         return v;
     }
 
-    public void setV(int v) {
+    public void setV(double v) {
         this.v = v;
     }
